homework-scripts/ex-4/tutorial-4-scripts/requirements.txt
greenlet
importlib-metadata
numpy
pandas
psycopg2-binary
python-dateutil
pytz
six
SQLAlchemy
typing-extensions
zipp